requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.5.0
djangorestframework==3.11.1
Markdown==3.2.2
markdown2==2.3.9
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
